core/trader.py
import os
import requests
from dotenv import load_dotenv

load_dotenv()

API_KEY = os.getenv("APCA_API_KEY_ID")
API_SECRET = os.getenv("APCA_API_SECRET_KEY")
BASE_URL = os.getenv("APCA_BASE_URL", "https://paper-api.alpaca.markets")

HEADERS = {
    "APCA-API-KEY-ID": API_KEY,
    "APCA-API-SECRET-KEY": API_SECRET,
    "Content-Type": "application/json"
}

def get_account():
    try:
        r = requests.get(f"{BASE_URL}/v2/account", headers=HEADERS)
        return r.json()
    except Exception as e:
        return {"error": str(e)}

def get_price(ticker):
    """Get latest price for a ticker using Alpaca market data"""
    try:
        url = f"https://data.alpaca.markets/v2/stocks/{ticker}/quotes/latest"
        r = requests.get(url, headers=HEADERS)
        data = r.json()
        # Use ask price as current price
        price = data.get("quote", {}).get("ap", None)
        if price:
            return float(price)
    except Exception:
        pass
    # Fallback mock prices if API fails
    mock_prices = {
        "AAPL": 174.50,
        "GOOGL": 142.30,
        "MSFT": 415.20,
        "TSLA": 175.80,
        "AMZN": 185.40
    }
    return mock_prices.get(ticker, 100.00)

def place_order(ticker, quantity, side):
    """Place a paper trade order on Alpaca"""
    try:
        payload = {
            "symbol": ticker,
            "qty": quantity,
            "side": side,
            "type": "market",
            "time_in_force": "day"
        }
        r = requests.post(f"{BASE_URL}/v2/orders", json=payload, headers=HEADERS)
        data = r.json()
        if "id" in data:
            return {
                "status": "filled",
                "order_id": data["id"],
                "symbol": ticker,
                "qty": quantity,
                "side": side
            }
        else:
            return {"status": "failed", "reason": data.get("message", "unknown error")}
    except Exception as e:
        # Fallback mock response
        return {
            "status": "filled (mock)",
            "order_id": "MOCK-001",
            "symbol": ticker,
            "qty": quantity,
            "side": side
        }

def get_positions():
    """Get current portfolio positions"""
    try:
        r = requests.get(f"{BASE_URL}/v2/positions", headers=HEADERS)
        return r.json()
    except Exception as e:
        return []