requirements.txt
alpaca-trade-api==3.2.0
python-dotenv==1.2.2
requests==2.32.5